Move more dis-like code from bytecode.py to dis.py
